+++ backend/src/main/java/com/example/focusapp/dto/GuestUserResponse.java
@@ -1,17 +1,17 @@
 package com.example.focusapp.dto;
 
 public class GuestUserResponse {
-    private Integer userId;
+    private Long userId;
     private String username;
     private String message;
 
-    public GuestUserResponse(Integer userId, String username, String message) {
+    public GuestUserResponse(Long userId, String username, String message) {
         this.userId = userId;
         this.username = username;
         this.message = message;
     }
 
-    public Integer getUserId() { return userId; }
+    public Long getUserId() { return userId; }
     public String getUsername() { return username; }
     public String getMessage() { return message; }
 }
