+++ backend/src/main/java/com/example/focusapp/repository/CharacterRepository.java
@@ -3,5 +3,5 @@ package com.example.focusapp.repository;
 import com.example.focusapp.entity.Character;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CharacterRepository extends JpaRepository<Character, Integer> {
+public interface CharacterRepository extends JpaRepository<Character, Long> {
 }
